feat: 거북이 정보 가져오기

+++ backend/auction-service/src/main/java/com/turtlecoin/auctionservice/domain/auction/dto/RegisterAuctionDTO.java
@@ -25,7 +25,7 @@ public class RegisterAuctionDTO {
     private Double minBid;
     private String content;
     private String title;
-    private Double weight;   // 거북이 무게 추가
+    private int weight;   // 거북이 무게 추가
     private Gender gender;   // 거북이 성별 추가
 
     public Auction toEntity() {
